append command

+++ cmd_decr.go
@@ -27,5 +27,5 @@ func DecrCommand(args []Value) Value {
 
 	intValue--
 	SETs[key] = strconv.Itoa(intValue)
-	return Value{typ: "integer", integer: int64(intValue)}
+	return Value{typ: "integer", integer: intValue}
 }
